+++ src/Handler/Method/BizumPayment.php
@@ -10,10 +10,10 @@ use Shopware\Core\Checkout\Customer\CustomerEntity;
 use Shopware\Core\Checkout\Order\OrderEntity;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
 
-class BancomatPayment extends PaymentHandler implements BankTransfer
+class BizumPayment extends PaymentHandler
 {
-    public const PAYMENT_METHOD_NAME = PaymentMethod::BANCOMATPAY;
-    public const PAYMENT_METHOD_DESCRIPTION = 'Bancomat Pay';
+    public const PAYMENT_METHOD_NAME = PaymentMethod::BIZUM;
+    public const PAYMENT_METHOD_DESCRIPTION = 'Bizum';
 
     protected string $paymentMethod = self::PAYMENT_METHOD_NAME;