+++ src/Service/PaymentMethodService.php
@@ -11,6 +11,7 @@ use Kiener\MolliePayments\Handler\Method\BanContactPayment;
 use Kiener\MolliePayments\Handler\Method\BankTransferPayment;
 use Kiener\MolliePayments\Handler\Method\BelfiusPayment;
 use Kiener\MolliePayments\Handler\Method\BilliePayment;
+use Kiener\MolliePayments\Handler\Method\BizumPayment;
 use Kiener\MolliePayments\Handler\Method\BlikPayment;
 use Kiener\MolliePayments\Handler\Method\CreditCardPayment;
 use Kiener\MolliePayments\Handler\Method\EpsPayment;
@@ -418,6 +419,7 @@ class PaymentMethodService
             MbWayPayment::class,
             MultibancoPayment::class,
             SwishPayment::class,
+            BizumPayment::class,
             // IngHomePayPayment::class, // not allowed anymore
             // DirectDebitPayment::class, // only allowed when updating subsriptions, aka => not allowed anymore
         ];